# Tide/Tide.Documentation/src/swagger-config.yaml
openapi: "3.0.0"
info:
  version: "0.0.1"
  title: "TIDE DOCUMENATION"
  description: "Tide C# & JS documentation"
servers:
  - url: "https://tidevendor.azurewebsites.net/tide/v1"
    description: "Tide Vendor API"
paths:
  # /test1:
  #   get:
  #     tags:
  #       - "onboarding"
  #     summary: "Test "
  #     description: "Returns string of something"
  #     produces:
  #       - "application/json"
  #     responses:
  #       "200":
  #         description: "Successful Operation"
  # /test/{name}/{noob}:
  #   get:
  #     tags:
  #       - "onboarding"
  #     summary: "Test endpoint"
  #     description: "Returns wether {name} if a {noob} or not"
  #     produces:
  #       - "application/json"
  #     parameters:
  #       - name: "name"
  #         in: "path"
  #         description: "The name of the person you're wondering about"
  #         required: true
  #         type: "string"
  #       - name: "noob"
  #         in: "path"
  #         description: "Is this person a noob or not"
  #         required: true
  #         type: "boolean"
  #     responses:
  #       "200":
  #         description: "Successful Operation"
  /GetUserNodes/{username}:
    get:
      tags:
        - "onboarding"
      summary: "Gets the nodes used by {username} on account creation"
      description: "Returns a an array of ork nodes"
      produces:
        - "application/json"
      parameters:
        - name: "username"
          in: "path"
          description: "Tideified username"
          required: true
          type: "string"
      responses:
        "200":
          description: "Successful Operation"
          schema:
            $ref: "#/definitions/TideResponse"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/TideResponse"
definitions:
  TideResponse:
    type: "object"
    properties:
      success:
        type: "boolean"
        default: false
      content:
        type: "string"
        default: null
      error:
        type: "string"
        default: null